修复已知问题 + 添加Notice模块 (#115)
* - Fix existing bugs
- Partially Finish notice part

* - Fix existing bugs
- Partially Finish notice part

* - Fix existing bugs

* - Fix existing bugs

* - Change Exception status code

* - Fix bugs

* Fix bug

* - Fix bug in GET /notice

* Remove teacher identity check

* - Fix bug in NoticeServiceImpl.java

* - Fix bug in NoticeServiceImpl.java

* - Fix bugs in NoticeServiceImpl.java

* - Fix bugs in select notice details

* Remove redundant exception throw

* Fix bug in PUT /notice

---------

+++ src/main/java/backend/service/impl/TeacherServiceImpl.java
@@ -208,10 +208,11 @@ public class TeacherServiceImpl implements TeacherService {
                 if (teacherProfile.getAvatar() != null) {
                     String avatar = targetUser.getAvatar();
                     if (avatar != null && !avatar.isEmpty()) {
-                        CompletableFuture.runAsync(() ->
-                                fileManager.deleteFile(avatar)
-                        );
-                        targetUser.setAvatar(fileManager.uploadBase64Image(avatar, targetUser));
+                        if(!targetUser.getAvatar().equals("default.png"))
+                            CompletableFuture.runAsync(() ->
+                                    fileManager.deleteFile(avatar)
+                            );
+                        targetUser.setAvatar(fileManager.uploadBase64Image(teacherProfile.getAvatar(), targetUser));
                     }
                 } else
                     targetUser.setAvatar(targetUser.getAvatar());
